backend/requirements.txt
fastapi
uvicorn
pandas
numpy
statsmodels
psycopg2-binary
sqlalchemy
python-dotenv
